places: Location class renamed to LatLng with doubles (BC break!)

+++ src/main/java/com/sygic/travel/sdk/trips/api/model/ApiTripItemResponse.kt
@@ -65,8 +65,8 @@ internal class ApiTripItemResponse(
 				) {
 					@JsonClass(generateAdapter = true)
 					internal class Location(
-						val lat: Float,
-						val lng: Float
+						val lat: Double,
+						val lng: Double
 					)
 				}
 			}
